requirements.txt
numpy
Pillow
colorama
cryptography
numba
pydub